Added a good pasta
Leeroy Jenkins created a strategy that revolved around trying to defeat your opponent in one turn without requiring any cards on the board. Fighting for board control and battles between minions make an overall game of Hearthstone more fun and compelling, but taking 20+ damage in one turn is not particularly fun or interactive.